Fix simulation integer params and error reporting (#38)

+++ src/api/routes/ingestion.ts
@@ -23,8 +23,8 @@ const uploadSchema = z.object({
 });
 
 const listQuerySchema = z.object({
-  page: z.coerce.number().min(1).default(1),
-  limit: z.coerce.number().min(1).max(100).default(20),
+  page: z.coerce.number().int().min(1).default(1),
+  limit: z.coerce.number().int().min(1).max(100).default(20),
   status: z.enum(['pending', 'processing', 'completed', 'failed']).optional(),
   sortBy: z.enum(['createdAt', 'name', 'status']).default('createdAt'),
   sortOrder: z.enum(['asc', 'desc']).default('desc'),
